refactor: update design (#644)
* import prettier

* Update Docusaurus dependencies to version 3.0.1

* Update package.json dependencies

* Add href prop to ProductCard component

* Refactor ProductCard component to include href prop

* Update SectionTitle component styling and custom.css

This commit updates the styling of the SectionTitle component in SectionTitle.tsx to support different modes (small and large). It also modifies the custom.css file to apply font and text color styles to the body element.

* Update translations and styles

* Add translation for Bitcoin custody message

* refactor: how it secures

* Update Chinese translations and add new messages

* Add background images to NetworkSecurity items

* Refactor code for improved performance and readability

* Refactor WalletSecurity component styling

* Update links in ServiceSecurity and WalletSecurity components

* revert package

* Add Prettier and Prettier Plugin for TailwindCSS to package.json

* Update custom.css and package.json

This commit updates the custom.css file to apply overflow-x-hidden to the root and html elements. It also updates the package.json file to use the latest versions of @docusaurus/theme-common, @unocss/reset, prism-react-renderer, react, react-dom, sonner, and tailwind-styled-components.

* Update Mixin ONE membership description